feat: migrate money storage from integer paise to decimal rupees
Every monetary field across appkit now stores decimal rupees directly
(e.g. 3499.00) instead of integer paise, eliminating the inconsistent
x100/รท100 conversions that caused display bugs (prices showing as
"35 lakhs" instead of "โน35,000"). Paise conversion now happens only at
the Razorpay/RazorpayX API boundaries.

- Renamed every *Paise/InPaise field to drop the suffix across products,
  lottery, orders, cart, stores, store-extensions, shipments, categories,
  and admin site-settings schemas.
- paiseSchema -> rupeesSchema (2dp-precision validator); deleted
  formatPaise() and two broken toRupees() helpers; added roundRupees().
- Core fee/GST/payout calculators now round in rupee-space.
- Rewrote ~50 seed data files to author decimal rupees directly.
- Adopted <PriceDisplay> (fixed to use the canonical formatCurrency) on
  ProductGrid/ProductDetailPageView instead of hand-rolled discount JSX.
- Added strict-zero audit-money-units.mjs to prevent regressions.

Also includes other pending working-tree changes (payment dispute/
fraud-reject/reupload flow, hard-ban reinstatement, checkout-value OTP,
payment-window timeout) that were already required by previously
committed code.

+++ src/features/catalogue/schemas/validation.ts
@@ -20,7 +20,7 @@ export const createCatalogueItemSchema = z.object({
   condition: productConditionSchema.optional(),
   categorySlugs: z.array(z.string()).optional(),
   brandSlug: z.string().optional(),
-  price: z.number().int().min(0).optional(),
+  price: z.number().min(0).optional(),
   quantity: z.number().int().min(1).default(1),
   visibility: z.enum(["public", "private"]).default("public"),
 });
